Fix content inset issues
This fixes content inset issues due to translucent navigation bars and tab bars. Prior, a `Spacer` was needed that calculated the correct size. This only worked for the navigation bar though, and not the tab bar on bottom.

This does something similar to a class in React Native (`RCTWrapperViewController`), and tries to calculate it automatically for any view (such as an `RCTScrollView`) that conforms to `RCTAutoInsetsProtocol`.

`Spacer` in theory should no longer be needed, but I have left it in for now in case anyone else should need it.