Cleanup warnings, implement some tab/tabbar props
This cleans up a number of warnings in the project, mostly unused variables and such, and cleans up comments that are no longer needed because what is commented has already been implemented.

It also adds a number of additional support in iOS for a number of different `UITabBar` and `UITabBarItem` props.