Add support for various modal presentation styles
This adds support for configuring the modal presentation style on a view presented modally in iOS. Full screen is the system default, which is cool, but not helpful when you want to show something as a form sheet in horizontally regular environments (iPhone plus sized in landscape, iPad), or as a popover.

This allows `modalPresentationStyle` to be passed as an option when presenting modally.