Implement additional nav bar props
In iOS, when a `UIViewController` is embedded inside a `UINavigationController`, setting the `preferredStatusBarStyle` (as done through the `statusBarStyle` prop in this library), has no effect, and `UINavigationController` returns either a light or dark status bar based on the bar style of its navigation bar. This will make things “work” in this case by allowing the developer to set if the barstyle is dark or not.